Initial commit for v1.0.
There's only one FullscreenActivity that permanently generates views and a touch listener that allows the user to interact with the spaceship. On startup the server time is asked. If it's been more than 6 hours since last stored-in-preferences crash, user can tap on Start button and play. When the rect of the spaceship intersects another rect (hitBoxes are smaller) the game ends and the crash time is stored.

There are lots of things to improve but this is just an initial version.

Also, a crash with log 'Cannot get dirty matrix' is happening to me sometimes but don't exactly know why (related to a task performed from the timer, but that is very generic information)